+++ test/helpers/globals.js
@@ -1,6 +1,0 @@
-const path = require("path")
-beforeAll(()=> {
-    console.log("proccess cwd ")
-    console.log(process.cwd())
-    jasmine.program = process.cwd()
-})